+++ apps/users/serializers/admin_withdrawals_analytics_serializers.py
@@ -16,3 +16,16 @@ class AdminWithdrawalTrendSerializer(serializers.Serializer[Any]):
     to_date = serializers.DateField()
     total = serializers.IntegerField()
     items = AdminWithdrawalTrendItemSerializer(many=True)
+
+class WithdrawalReasonPercentageItemSerializer(serializers.Serializer[Any]):
+    reason = serializers.CharField()
+    reason_label = serializers.CharField()
+    count = serializers.IntegerField()
+    percentage = serializers.FloatField()
+
+
+class WithdrawalReasonPercentageSerializer(serializers.Serializer[Any]):
+    from_date = serializers.DateField()
+    to_date = serializers.DateField()
+    total = serializers.IntegerField()
+    items = WithdrawalReasonPercentageItemSerializer(many=True)
